aprs-to-ws/requirements.txt
pyserial
websocket-server
aprspy
bitstring
